Fix for loss of precision when using average_hash with a hash_size != 8.
In order to convert the full bit string without loss of precision into a hex string, the max padded length must be hash_size ** 2 / 4 (bits_per_hex_digit). In the previous implementation, a hash_size * 2 format string was generated, which would not add enough (or any) padded zeros and would result in hashes of unequal length.